Mockup of alg2
Tests are failing every once in a while so I wanted to see if there was
any pattern for which 2 algorighms get the same answer. Jaerod, I know
you had dibs on this algorithm, so feel free to delete the file I
added and replace it with your own work when you have time. Or better
yet, just skip straight to debugging the broken tests :)